requirements.txt
# 🎯 سیستم استخراج هوشمند اسناد گمرکی
# Dependencies List

# GUI Framework
tkinter>=8.6.0  # Built-in with Python 3.x
customtkinter>=5.2.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# PDF Processing
PyMuPDF>=1.23.0

# OCR Engine
easyocr>=1.7.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Other Utilities
pathlib>=1.0.1  # Built-in with Python 3.4+
logging>=0.4.9.6  # Built-in
json>=2.0.9  # Built-in
re>=2.2.1  # Built-in
datetime>=4.3  # Built-in
threading>=1.0  # Built-in
queue>=1.0  # Built-in
hashlib>=20081119  # Built-in
pickle>=1.0  # Built-in

# Optional but recommended
matplotlib>=3.7.0  # برای نمودارها (اختیاری)
seaborn>=0.12.0  # برای تصویرسازی بهتر (اختیاری)
